resue proxy handling
reuse exporter actually detectes whether a proxy is ciffable now
and either converts it to a symbol call if it is,
or just steamrolls it.
Added Reuse.stat.steamrolls to keep track of how many
compos were steamrolled
and a test that makes sure
no steamroll is made when converting to symbol call
is possible.